# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.passwordChange:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  handlers.passwordReset:
    properties:
      new_password:
        type: string
    type: object
...
  handlers.userCreate:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  handlers.userUpdate:
    properties:
      is_active:
        type: boolean
      role:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: 多USB 4G Modem 管理系统 API
  title: SimNexus API
  version: "1.0"
paths:
...
  /api/v1/users/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.userCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    patch:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.userUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改用户信息
      tags:
      - 用户管理
  /api/v1/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.passwordReset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 管理员重置用户密码
      tags:
      - 用户管理
  /api/v1/users/me/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 旧密码和新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.passwordChange'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改当前用户密码
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"